+++ hardpy/__init__.py
@@ -1,7 +1,6 @@
 # Copyright (c) 2024 Everypin
 # GNU General Public License v3.0 (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
 
-from hardpy.pytest_hardpy.plugin import HardpyPlugin
 from hardpy.pytest_hardpy.result import CouchdbLoader
 from hardpy.pytest_hardpy.utils import DuplicateSerialNumberError
 from hardpy.pytest_hardpy.result.couchdb_config import CouchdbConfig
@@ -18,7 +17,6 @@ from hardpy.pytest_hardpy.pytest_call import (
 )
 
 __all__ = [
-    "HardpyPlugin",
     "CouchdbLoader",
     "CouchdbConfig",
     "DuplicateSerialNumberError",